refactor(build): split native dependency fixtures
